+++ src/components/sectionedForm/index.ts
@@ -1,6 +1,5 @@
 export * from './SectionedFormSections'
 export * from './SectionedFormFooter'
-export * from './DefaultSectionedFormFooter'
 export * from './DefaultSectionedFormSidebar'
 export * from './SectionedFormLayout'
 export * from './SectionedFormErrorNotice'