+++ src/components/standardForm/StandardFormActions.tsx
@@ -1,10 +1,15 @@
 import { Button, ButtonStrip, CircularLoader } from '@dhis2/ui'
 import React from 'react'
+import { To } from 'react-router-dom'
+import { LinkButton } from '../LinkButton'
 import classes from './StandardFormActions.module.css'
 
 function LoadingIcon() {
     return (
-        <span className={classes.submitButtonLoadingIcon}>
+        <span
+            className={classes.submitButtonLoadingIcon}
+            data-test="loading-icon"
+        >
             <CircularLoader small />
         </span>
     )
@@ -14,28 +19,37 @@ export function StandardFormActions({
     cancelLabel,
     submitLabel,
     submitting,
-    onCancelClick,
+    onSubmitClick,
+    cancelTo,
 }: {
     cancelLabel: string
     submitLabel: string
     submitting: boolean
-    onCancelClick: () => void
+    onSubmitClick: () => void
+    cancelTo?: To
 }) {
     return (
         <ButtonStrip>
-            <Button primary small disabled={submitting} type="submit">
+            <Button
+                primary
+                small
+                disabled={submitting}
+                type="submit"
+                onClick={onSubmitClick}
+                dataTest="form-submit-button"
+            >
                 {submitting && <LoadingIcon />}
                 {submitLabel}
             </Button>
-
-            <Button
+            <LinkButton
                 secondary
                 small
                 disabled={submitting}
-                onClick={onCancelClick}
+                to={cancelTo}
+                dataTest="form-cancel-link"
             >
                 {cancelLabel}
-            </Button>
+            </LinkButton>
         </ButtonStrip>
     )
 }