+++ src/pages/categoryOptionCombos/List.spec.tsx
@@ -1,14 +1,53 @@
+import categoryOptionCombosSchema from '../../__mocks__/schema/categoryOptionCombosSchema.json'
+import { SECTIONS_MAP } from '../../lib'
+import {
+    testCategoryCombo,
+    testCategoryOptionCombo,
+} from '../../testUtils/builders'
 import {
     generateDefaultListFiltersTests,
     generateDefaultListItemsTests,
     generateDefaultListMultiActionsTests,
-    generateDefaultListRowActionsTests,
 } from '../defaultListTests'
+import { Component as CategoryOptionComboList } from './List'
 
-const componentName = 'Category option combo'
-generateDefaultListItemsTests({ componentName })
-generateDefaultListFiltersTests({ componentName })
-generateDefaultListMultiActionsTests({ componentName })
+const section = SECTIONS_MAP.categoryOptionCombo
+const mockSchema = categoryOptionCombosSchema
+const ComponentToTest = CategoryOptionComboList
+const generateRandomElement = testCategoryOptionCombo
+const customData = {
+    categoryCombos: (type: any, params: any) => {
+        if (type === 'read') {
+            return {
+                categoryCombos: [
+                    testCategoryCombo(),
+                    testCategoryCombo(),
+                    testCategoryCombo(),
+                ],
+                pager: { page: 1, total: 3, pageSize: 20, pageCount: 1 },
+            }
+        }
+    },
+    categoryOptions: (type: any, params: any) => {
+        if (type === 'read') {
+            return {
+                categoryOptions: [],
+                pager: { page: 1, total: 0, pageSize: 20, pageCount: 1 },
+            }
+        }
+    },
+}
+
+const testConfig = {
+    section,
+    mockSchema,
+    ComponentToTest,
+    generateRandomElement,
+    customData,
+}
+generateDefaultListItemsTests(testConfig)
+generateDefaultListFiltersTests(testConfig)
+generateDefaultListMultiActionsTests(testConfig)
 
 xdescribe('Category option combo additional tests', () => {
     it('should filter by category option', () => {})