+++ src/pages/categoryOptions/New.spec.tsx
@@ -1,7 +1,3 @@
-import { generateDefaultAddFormTests } from '../defaultFormTests'
-
-generateDefaultAddFormTests({ componentName: 'Category option' })
-
 xdescribe('Category option add form additional tests', () => {
     it('contain all needed field', () => {})
     it('should not submit when required values are missing', () => {})