+++ src/pages/dataElementGroupSets/form/dataElementGroupSetSchema.ts
@@ -1,11 +1,14 @@
 import { z } from 'zod'
+import { modelFormSchemas } from '../../../lib'
 
-export const dataElementGroupSetSchema = z
-    .object({
-        name: z.string().trim(),
-        shortName: z.string().trim(),
-        code: z.string().trim(),
-        description: z.string().trim(),
-        dataElements: z.array(z.object({ id: z.string() })),
-    })
-    .partial()
+const { identifiable } = modelFormSchemas
+export const dataElementGroupSetSchema = identifiable.merge(
+    z
+        .object({
+            shortName: z.string().trim(),
+            code: z.string().trim(),
+            description: z.string().trim(),
+            dataElements: z.array(z.object({ id: z.string() })),
+        })
+        .partial()
+)