+++ src/pages/indicatorTypes/form/index.ts
@@ -1,1 +1,1 @@
-export { IndicatorSchema, validate } from './IndicatorTypesSchema'
+export { IndicatorTypeFormSchema, validate } from './IndicatorTypesSchema'